Add wind scale conversion function for displaying wind strength
The code changes in this commit add a new function, convertMphToWindScale, to convert wind speed in miles per hour (mph) to a descriptive wind scale. This function is used in the WeatherComponent in Home.vue to display the average wind strength over the last 10 minutes. By providing a clear and meaningful wind scale, users can easily understand the intensity of the wind.